Bug fix in weak checksum implementation
Previously, the start and end index of a block
weren't taken into account when calculating the
weak checksum for a given block. This is incorrect as
the weak checksum described in the rsync paper uses
indices.
Instead of using actual index values, everytime the
start index was set to be zero and the end index was set
to be the block size.
This has been fixed now. The actual index values are taken
into account.