Implement Rolling Checksum Functionality
Implemented the rolling checksum function, the
one described in the rsync paper.
Also edited the return value of the weakChecksum
function. It now returns all the components of the
checksum(a, b, and s) as this is of use in other
functions, like calculating the rolling checksum.
The components of the weak checksum are defined in the
rsync algorithm paper.