Implement functionality to create delta file
Implemented the logic to create the delta file.
It is assumed that there won't be more than 2^32 block indices.
It is also assumed that the length of the literals batched under
one literal command won't exceed 2^32.
For the POC and in general, these assumptions won't pose an issue.
In a more robust implementation, these need to be standardized.