Empty buffer before ending the delta file
If the entire file differs, then the entire file
is copied. In this case, the buffer gets filled with
the contents of the entire file and must be written to the
file before the end command is written.
Previously, this case was ignored and the contents of the buffer
were not written to the file before writing the end command.