Remove code which reads a block outside the loop
Removed code which read a block from the in-file
outside the loop. This code was present when testing
the program and shouldn't have been present in the
code.
This leads to problems as a block is already read and
nothing is done. The checksum calculations aren't performed
on it and hence this code has to be removed.