Add files required to upload package to pypi
Added the files required to upload packages to pypi.
The previous releases(0.1.0, 0.1.1) had installation issues associated with
them. "functools" was listed as an installation requirement, which isn't
necessary as its part of the core python packages.
Removing this requirement from the setup.cfg solves the issue when trying
to install the rdiff package using pip.